+++ wiki/scripts/dir/helpers.js
@@ -1,0 +1,3 @@
+export function headingToAnchor(heading) {
+    return encodeURI(heading.split(":")[0].split(" ")[0].replace("?", ""));
+}