+++ wiki/scripts/dir/markdown.js
@@ -1,0 +1,6 @@
+import md from "./markdown/index.js";
+import * as __ from "./markdown/anchor.js";
+import * as ___ from "./markdown/section.js";
+__;
+___;
+export default md;