+++ wiki/scripts/dir/markdown/index.js
@@ -1,0 +1,5 @@
+const md = window.markdownit({
+    typographer: true,
+});
+export const proxy = (tokens, idx, options, env, self) => self.renderToken(tokens, idx, options);
+export default md;