+++ wiki/scripts/src/markdown.ts
@@ -1,0 +1,12 @@
+// Edit Markdown rules https://github.com/markdown-it/markdown-it/blob/master/docs/examples/renderer_rules.md
+// Use https://markdown-it.github.io/
+
+import md from "./markdown/index.js"
+import * as __ from "./markdown/anchor.js"
+import * as ___ from "./markdown/section.js"
+// import * as ____ from "./markdown/linkify.js";
+__
+___
+// ____
+
+export default md